Move ConfigRunner to Commands.ConfigCommand namespace
