+++ Unit4.Automation.Tests/ConfigRunnerTests.cs
@@ -3,6 +3,7 @@ using Unit4.Automation.Model;
 using Unit4.Automation;
 using NUnit.Framework;
 using Unit4.Automation.Tests.Helpers;
+using Unit4.Automation.Commands.ConfigCommand;
 
 namespace Unit4.Automation.Tests
 {
